+++ easypost/beta/__init__.py
@@ -1,2 +1,3 @@
 # flake8: noqa
+from easypost.beta.rate import Rate
 from easypost.beta.referral import Referral