+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # CHANGELOG
 
-## Next Release
+## v7.10.0 (2023-02-17)
 
-- Adds `retrieve_stateless_rates` function to pull stateless rates when shipment data is provided
+- Adds beta `retrieve_stateless_rates` function
 - Adds `get_lowest_stateless_rate` function to filter the lowest stateless rate
 
 ## v7.9.0 (2023-01-18)